third_party/WebKit/LayoutTests/fast/dom/custom/callback-timing-expected.txt

Ensure that custom element callback fires before DOMContentLoaded event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS eventlog[0] is 'created x-foo'
PASS eventlog[1] is 'DOMContentLoaded event'
PASS eventlog[2] is 'load event'
